Mergesh

+++ src/main/java/com/project3/placestation/product/controller/ProductController.java
@@ -2,9 +2,9 @@ package com.project3.placestation.product.controller;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,21 +13,26 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import com.project3.placestation.biz.handler.exception.CustomLoginRestfulException;
 import com.project3.placestation.biz.model.dto.ResProductDto;
+import com.project3.placestation.biz.model.util.BizDefine;
+import com.project3.placestation.payment.model.dto.PaymentMemberDto;
 import com.project3.placestation.product.dto.ProdReviewDto;
 import com.project3.placestation.product.dto.ProdWishListDto;
 import com.project3.placestation.product.dto.ProductInvalidDateDto;
+import com.project3.placestation.product.dto.ResProductViewDto;
 import com.project3.placestation.repository.entity.AdditionExplanation;
-import com.project3.placestation.repository.entity.ProdReview;
-import com.project3.placestation.repository.entity.Product;
+import com.project3.placestation.repository.entity.Member;
 import com.project3.placestation.repository.interfaces.ProductRepository;
 import com.project3.placestation.service.AddtionExplanationService;
 import com.project3.placestation.service.AdminProdHistoryService;
+import com.project3.placestation.service.MemberService;
 import com.project3.placestation.service.ProdReviewService;
 import com.project3.placestation.service.ProdWishListService;
 import com.project3.placestation.service.ProductService;
 import com.project3.placestation.service.ProductViewService;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -49,69 +54,101 @@ public class ProductController {
 	AdminProdHistoryService adminProdHistoryService;
 	@Autowired
 	ProductViewService productViewService;
+	@Autowired
+	MemberService memberService;
+	
+	@Autowired
+	HttpSession httpSession;
 
 	//http://localhost:80/productDetail?prod_no=
 	@GetMapping("/productDetail")
 	public String productDetail(@RequestParam("prod_no") Integer prodNo,
-            @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
-            Model model) {
-		log.debug("상품 상세 페이지 - 상품번호: {}", prodNo);
-		
-	    // 상품 상세 페이지 접속 시 조회수 증가
-	    productViewService.increaseProductViews(prodNo);
+	        @RequestParam(name = "pageNo", defaultValue = "1") Integer pageNo,
+	        Model model, HttpSession session) throws Exception {
+	    
+	    // 조회수 세션 저장 (새로고침 조회수 증가 x)
+	    Boolean increaseProductViews = (Boolean) session.getAttribute("increaseProductViews");
+	    if (increaseProductViews == null || increaseProductViews) {
+	        productViewService.increaseProductViews(prodNo);
+	        session.setAttribute("increaseProductViews", false);
+	    }
 	    // 상품 조회수 가져오기
-	    int currentViews = productViewService.getProductViews(prodNo);
-		// 상품 번호로 조회
-		ResProductDto product = productService.findById(prodNo);
-		// 상품의 리뷰 개수 조회
-        Integer reviewCount = prodReviewService.getCountReview(prodNo);
-        
+	    ResProductViewDto currentViews = productViewService.getProductViews(prodNo);
+	    // 상품 번호로 조회
+	    ResProductDto product = productService.findById(prodNo);
+	    // 상품의 리뷰 개수 조회
+	    Integer reviewCount = prodReviewService.getCountReview(prodNo);
+	    
 	    // 페이지당 리뷰 수 설정
 	    int reviewsPerPage = 5;
 	    // 총 페이지 수 계산
 	    int totalPage = (int) Math.ceil((double) reviewCount / reviewsPerPage);
 
-		// 상품 번호로 리뷰 조회
-		List<ProductInvalidDateDto> invalidDate = adminProdHistoryService.findProductInvalidByProdNo(prodNo, "");
-		
+	    // 상품 번호로 리뷰 조회
+	    List<ProductInvalidDateDto> invalidDate = adminProdHistoryService.findProductInvalidByProdNo(prodNo, "");
+	    
 	    // 리뷰 목록을 페이징하여 조회
 	    List<ProdReviewDto> reviewProdNo = prodReviewService.findByRevProdNoPaged(prodNo, pageNo * (reviewsPerPage - 4), reviewsPerPage);
-	    log.debug("페이지번호 : " + pageNo );
+	    // 리뷰 목록의 작성자 정보 설정
+	    for (ProdReviewDto review : reviewProdNo) {
+	        PaymentMemberDto reviewer = memberService.findMemberById(review.getUserNo());
+	        if (reviewer != null) {
+	            review.setUserName(reviewer.getUserName());
+	        }
+	    }
+	    // 상품의 사업자명 설정
+	    PaymentMemberDto writer = memberService.findMemberById(product.getProdWriterNo());
+	    if (writer != null) {
+	        product.setUserName(writer.getUserName());
+	    }
 	    // 부가 설명 이미지
 	    List<AdditionExplanation> additionExplanations = addtionExplanationService.findAll();
 	    
-        // 상품의 찜 개수 조회
-		Integer wishlistCount = prodWishListService.getCountWishlist(prodNo);
-        Double avgStar = prodReviewService.getAvgStar(prodNo);
-        
-        // 부가 설명 이미지 뽑기
-        List<AdditionExplanation> list = new ArrayList<>();
-        for(String i : product.getAdditionExplanation()) {
-        	list.add(additionExplanations.get(Integer.valueOf(i) - 1));
-        }
-        
-        log.info(list.toString());
+	    // 현재 로그인한 사용자의 정보 가져오기
+	    Member member = (Member) session.getAttribute("member");
+	    if (member != null) {
+		    // 로그인한 사용자의 찜 목록에 해당 상품이 있는지 확인
+	        boolean isProductInWishlist = prodWishListService.isProductInWishlist(prodNo, member.getUserno());
+	        model.addAttribute("isProductInWishlist", isProductInWishlist);
+	    }
+	    
 
-		log.info(invalidDate.toString());
-		model.addAttribute("product", product);
-		model.addAttribute("reviewProdNo", reviewProdNo);
-		model.addAttribute("invalidDate", invalidDate);
-		model.addAttribute("wishlistCount", wishlistCount);
-		model.addAttribute("additionExplanations", list);
-		model.addAttribute("reviewCount", reviewCount);
-		model.addAttribute("avgStar", avgStar);
+	    // 상품의 찜 개수 조회
+	    Integer wishlistCount = prodWishListService.getCountWishlist(prodNo);
+	    Double avgStar = prodReviewService.getAvgStar(prodNo);
+	    
+	    // 부가 설명 이미지 뽑기
+	    List<AdditionExplanation> list = new ArrayList<>();
+	    for(String i : product.getAdditionExplanation()) {
+	        list.add(additionExplanations.get(Integer.valueOf(i) - 1));
+	    }
+	    
+	    log.info(list.toString());
+
+	    log.info(invalidDate.toString());
+	    model.addAttribute("product", product);
+	    model.addAttribute("reviewProdNo", reviewProdNo);
+	    model.addAttribute("invalidDate", invalidDate);
+	    model.addAttribute("wishlistCount", wishlistCount);
+	    model.addAttribute("additionExplanations", list);
+	    model.addAttribute("reviewCount", reviewCount);
+	    model.addAttribute("avgStar", avgStar);
 	    model.addAttribute("pageNo", pageNo); // 현재 페이지 번호 추가
 	    model.addAttribute("totalPage", totalPage); // 총 페이지 수 추가
-	    model.addAttribute("currentViews", currentViews);
+	    model.addAttribute("currentViews", currentViews.getProdViews());
 
-		return "product/productDetail";
+	    return "product/productDetail";
 	}
 
 	// 답글 작성
 	@PostMapping("/saveReview")
 	public String saveReview(ProdReviewDto dto, @RequestParam("prodNo") Integer prodNo) {
 		dto.setProdNo(prodNo);
-
+		
+	    // 유효성 검사
+	    if (dto.getUserNo() == null) {
+	        throw new CustomLoginRestfulException("로그인 후 작성해주시기 바랍니다.", HttpStatus.INTERNAL_SERVER_ERROR);
+	    }
 		prodReviewService.saveReview(dto);
 
 		return "redirect:/product/productDetail?prod_no=" + prodNo;
@@ -119,14 +156,24 @@ public class ProductController {
 
 	// 리뷰 작성
 	@PostMapping("/addReview")
-	public String addReview(ProdReviewDto dto, @RequestParam("prodNo") Integer prodNo) {
+	public String addReview(ProdReviewDto dto, @RequestParam("prodNo") Integer prodNo, Model model, HttpSession session) {
 		dto.setProdNo(prodNo);
 
-		prodReviewService.addReview(dto);
+		
+		// 멤버 받기
+		Member member = (Member) session.getAttribute("member"); 
+		if(member == null) {
+			throw new CustomLoginRestfulException(BizDefine.ACCOUNT_IS_NONE, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
 
+	    prodReviewService.addReview(dto);
+
+		PaymentMemberDto memberDto = memberService.findMemberById(member.getUserno());
+		
+	    model.addAttribute("member", memberDto);
+	    
 		return "redirect:/product/productDetail?prod_no=" + prodNo;
 	}
-
 
 	// 리뷰 삭제
 	@PostMapping("/deleteReview/{prodRevNo}")
@@ -138,45 +185,31 @@ public class ProductController {
 	// 찜(상품 좋아요) 하기
 	@PostMapping("/addWishlist")
 	public String addWishlist(ProdWishListDto dto, @RequestParam("prodNo") Integer prodNo) {
+		// 멤버 받기
+		Member member = (Member) httpSession.getAttribute("member"); 
+		if(member == null) {
+			throw new CustomLoginRestfulException(BizDefine.ACCOUNT_IS_NONE, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
         dto.setProdNo(prodNo);
+        dto.setUserNo(member.getUserno());
         prodWishListService.addWishList(dto);
 		return "redirect:/product/productDetail?prod_no=" + prodNo;
 	}
-	
 
 	// 찜 삭제
     @PostMapping("/deleteWishlist")
     public String deleteWishlist(ProdWishListDto dto, @RequestParam("prodNo") Integer prodNo) {
+    	
+		// 멤버 받기
+		Member member = (Member) httpSession.getAttribute("member"); 
+		if(member == null) {
+			throw new CustomLoginRestfulException(BizDefine.ACCOUNT_IS_NONE, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
         dto.setProdNo(prodNo);
+        dto.setUserNo(member.getUserno());
         prodWishListService.deleteWishList(dto);
         return "redirect:/product/productDetail?prod_no=" + prodNo;
     }
-    
-    //http://localhost:80/product/main
-	@GetMapping("/main")
-	public String mainindex(Model model) {
-		log.debug("메인 페이지!");
-		// 상품 전체 리스트 조회
-		List<ResProductDto> products = productService.findAll();
-
-		// 전체 상품 조회 4개
-
-		List<ResProductDto> topProducts = products.stream().limit(8).collect(Collectors.toList());
-		// 리뷰 많은 상품
-		List<ProdReview> productsRev = productRepository.findAllByRev();
-		// 평점 높은 상품
-		List<Product> productsStar = productRepository.findAllByStar();
-		// 최신 상품
-		List<Product> productsStart = productRepository.findAllByStart();
-
-		model.addAttribute("products", topProducts);
-		model.addAttribute("productsRev", productsRev);
-		model.addAttribute("productsStar", productsStar);
-		model.addAttribute("productsStart", productsStart);
-
-		return "product/main";
-	}
-
 
 
 }
